+++ src/InputFilterPluginManagerFactory.php
@@ -8,6 +8,7 @@
 namespace Zend\InputFilter;
 
 use Interop\Container\ContainerInterface;
+use Zend\ServiceManager\Config;
 use Zend\ServiceManager\FactoryInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
 
@@ -27,7 +28,30 @@ class InputFilterPluginManagerFactory implements FactoryInterface
      */
     public function __invoke(ContainerInterface $container, $name, array $options = null)
     {
-        return new InputFilterPluginManager($container, $options ?: []);
+        $pluginManager = new InputFilterPluginManager($container, $options ?: []);
+
+        // If this is in a zend-mvc application, the ServiceListener will inject
+        // merged configuration during bootstrap.
+        if ($container->has('ServiceListener')) {
+            return $pluginManager;
+        }
+
+        // If we do not have a config service, nothing more to do
+        if (! $container->has('config')) {
+            return $pluginManager;
+        }
+
+        $config = $container->get('config');
+
+        // If we do not have input_filters configuration, nothing more to do
+        if (! isset($config['input_filters']) || ! is_array($config['input_filters'])) {
+            return $pluginManager;
+        }
+
+        // Wire service configuration for input_filters
+        (new Config($config['input_filters']))->configureServiceManager($pluginManager);
+
+        return $pluginManager;
     }
 
     /**